Refactor binance apps
