+++ src/exchanges/binance/mod.rs
@@ -1,5 +1,15 @@
+pub mod account;
 pub mod auth;
-mod client;
-mod types;
+pub mod client;
+pub mod converters;
+pub mod market_data;
+pub mod trading;
+pub mod types;
 
+// Re-export main types for easier importing
 pub use client::BinanceConnector;
+pub use types::{
+    BinanceAccountInfo, BinanceBalance, BinanceExchangeInfo, BinanceFilter, BinanceKlineData,
+    BinanceMarket, BinanceOrderRequest, BinanceOrderResponse, BinanceWebSocketKline,
+    BinanceWebSocketOrderBook, BinanceWebSocketTicker, BinanceWebSocketTrade,
+};
