+++ src/exchanges/binance_perp/mod.rs
@@ -1,4 +1,15 @@
+pub mod account;
 pub mod client;
+pub mod converters;
+pub mod market_data;
+pub mod trading;
 pub mod types;
 
+// Re-export main types for easier importing
 pub use client::BinancePerpConnector;
+pub use types::{
+    BinancePerpBalance, BinancePerpExchangeInfo, BinancePerpFilter, BinancePerpKlineData,
+    BinancePerpMarket, BinancePerpOrderRequest, BinancePerpOrderResponse, BinancePerpPosition,
+    BinancePerpWebSocketKline, BinancePerpWebSocketOrderBook, BinancePerpWebSocketTicker,
+    BinancePerpWebSocketTrade,
+};
